개별 조회 템플릿 완료 & 삭제 기능 구현
