+++ docs/200324.md
@@ -1,0 +1,21 @@
+# **ExpressJS**
+* [API](http://expressjs.com/ko/4x/api.html#express)
+## (response) redirect
+```
+res.redirect('/');
+// 어떤 데이터를 보낼지는 생각할 필요 없음 (redirect 시킨 라우터에서 자동 처리)
+```
+## (request) header 관련 데이터
+```
+req.headers.{property}
+// html 안의 form 값이 위변조 가능하다면 post 자체에서 편집이 어려운 값을 받는 것도 고려
+// but 보내는 url도 충분히 postman 같은 걸로 변경 가능할 것으로 예상되니, 궁극적으로는 session 도입해야함
+```
+# 정규표현식 in JavaScript
+## match
+* MDN ([참조](https://developer.mozilla.org/ko/docs/Web/JavaScript/Reference/Global_Objects/String/match))
+```
+const id = String(req.headers.referer).match(/review\/(\d+)/)[1]
+// match 메소드를 쓰면 결과값이 없으면 null을 결과값이 있으면 Array return
+// Array엔 전체 검색된 텍스트, (있다면 그룹들 ...), index, input, groups 순으로 존재
+```
