+++ views/pages/detail.ejs
@@ -1,50 +1,87 @@
-<!-- 가게 상세 모달 -->
-<div class="modal fade" id="expandStoreModal" tabindex="-1" role="dialog" aria-labelledby="expandStoreModalLabel"
-    aria-hidden="true">
-    <div class="modal-dialog modal-lg" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <div>
-                    <h5 class="modal-title" id="expandStoreModalLabel">가게 상세 정보</h5>
-                </div>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                    <span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-            <div class="modal-body container-fluid">
-                <div class="display-4 modal-title h3 font-weight-bold" id="expandStoreName">가게 이름</div>
-                <div id="expandStoreAdress" class="lead">가게 주소
-                </div>
-                <small id="expandStoreCategory" class="font-italic">카테고리</small>
-                <div class="row">
-                    <div class="col-lg-6 p-3 d-flex flex-column">
-                        <div class="h4">이미지</div>
-                        <div class="d-flex align-items-center flex-fill">
-                            <img id="expandStoreImg" class="img-fluid border rounded" src="img/meal.png">
-                        </div>
-                    </div>
-                    <div class="col-lg-6 p-3">
-                        <div class="h4">지도</div>
-                        <div id="expandStoreMap" class="mx-auto border rounded">API</div>
-                    </div>
-                    <div class="col-2 h4 p-3">별점</div>
-                    <div class="col m-3 storeStar" id="expandStoreStar">별별별별별</div>
-                    <div class="w-100"></div>
-                    <div class="col p-3">
-                        <div class="h4">리뷰</div>
-                        <div id="expandStoreReview" class="p-3"></div>
-                    </div>
-                    <div class="w-100"></div>
-                    <div class="col d-flex justify-content-start">
-                        <button id="watchLarge" class="mr-1 btn btn-outline-primary">큰 지도 보기</button>
-                        <button id="watchDetail" class="mr-1 btn btn-outline-primary">상세 페이지 보기</button>
-                        <button id="findWay" class="btn btn-outline-primary">길찾기</button>
-                    </div>
-                </div>
-            </div>
-            <div id="expandStoreTimeStamp" class="modal-footer text-right small font-italic">
+<!doctype html>
+<html lang="ko">
 
+<head>
+    <%- include('../partials/header', {title: title}); %>
+</head>
+
+<body>
+    <%- include('../partials/nav'); %>
+    <section class="container">
+        <div class="container-fluid border rounded">
+            <header>
+                <h1 class="display-4 font-weight-bold" id="storeName"><%= review.name %></h1>
+                <p class="lead"><%= review.address %></p>
+                <%# 카카오 API를 통해 비동기 작업 %>
+                <small id="storeCategory" class="font-italic">카테고리</small>
+                <div class="text-right">
+                    <small>
+                        작성시간 : <span id="storeCreated"
+                            class="font-italic"><%= review.createdAt.toLocaleString() %></span>
+                    </small>
+                    <% if(review.createdAt.toString() != review.updatedAt.toString()) { %>
+                    <br>
+                    <small>
+                        수정시간 : <span id="storeUpdated"
+                            class="font-italic"><%= review.updatedAt.toLocaleString() %></span>
+                    </small>
+                    <% } %>
+                    </p>
+                </div>
+            </header>
+            <section class="my-3">
+                <h4>사진</h4>
+                <figure class="d-flex justify-content-center p-3">
+                    <img id="storeImage" class="img-fluid border rounded" src=<%# 차후에 test 옵션은 삭제 %>
+                        <% if (review.address == 'test' || review.address == '') { %> <%= "/img/meal.png" %>
+                        <% } else { %> <%= "review.image" %> <% } %> class="img-fluid" alt="리뷰 이미지">
+                </figure>
+                <%# 카카오 API를 통해 비동기 작업 %>
+                <h4>지도</h4>
+                <figure class="p-3">
+                    <div id="storeMap" class="mx-auto border rounded">API</div>
+                </figure>
+                <h4>별점</h4>
+                <figure class="p-3">
+                    <div class="storeStar" id="storeStar">☆☆☆☆☆</div>
+                </figure>
+                <h4>리뷰</h4>
+                <article class="p-3">
+                    <p><%= review.comments %></p>
+                </article>
+            </section>
+            <div class="d-flex justify-content-center my-3">
+                <button id="watchLarge" class="mr-2 btn btn-outline-primary">큰 지도 보기</button>
+                <button id="watchDetail" class="mr-2 btn btn-outline-primary">상세 페이지 보기</button>
+                <button id="findWay" class="mr-2 btn btn-outline-primary">길찾기</button>
+                <button class="btn btn-outline-danger" data-toggle="modal" data-target="#deleteModal">삭제하기</button>
+            </div>
+        </div>
+    </section>
+    <div id="deleteModal" class="modal" tabindex="-1" role="dialog">
+        <div class="modal-dialog" role="document">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body">
+                    <p>정말 <b>삭제</b>하시겠습니까?</p>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">닫기</button>
+                    <form method="POST" action="/delete">
+                        <button type="submit" class="btn btn-danger">삭제하기</button>
+                    </form>
+                </div>
             </div>
         </div>
     </div>
-</div>
+    <%- include('../partials/scripts'); %>
+    <script>
+        drawStar(<%= review.rate %>, document.querySelector('#storeStar'));
+    </script>
+</body>
+
+</html>
