+++ views/partials/header.ejs
@@ -4,6 +4,6 @@
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 <%# Bootstrap CSS %>
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
-<link rel="stylesheet" href="./css/main.css">
+<link rel="stylesheet" href="/css/main.css">
 <%# 전달 받은 title을 표시 %>
 <title><%= title %></title>
